acorda-backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0
PyJWT>=2.8,<3.0

# Async Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.5,<3.0

# Payments - Mercado Pago
mercadopago>=2.2,<3.0

# Email
django-anymail[sendgrid]>=10.2,<11.0
# Alternative: django-anymail[amazon_ses] or django-anymail[postmark]

# Storage (S3)
django-storages>=1.14,<2.0
boto3>=1.34,<2.0

# Security
python-decouple>=3.8,<4.0
argon2-cffi>=23.1,<24.0

# Utils
python-dateutil>=2.8,<3.0
uuid>=1.30

# Production
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0

# Development
django-debug-toolbar>=4.2,<5.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
